@@ -13,9 +13,9 @@ dependencies:
   get: ^4.6.6
   dio: ^5.7.0
   google_fonts: ^6.2.1
-  font_awesome_flutter: ^10.7.0
+  font_awesome_flutter: ^10.8.0
   flutter_svg_provider: ^1.0.7
-  path_provider: ^2.1.4
+  path_provider: ^2.1.5
   flutter_screenutil: ^5.9.3
 
 #  readmore: ^3.0.0
@@ -25,19 +25,19 @@ dependencies:
   image_cropper: ^8.0.2
 
   select2dot1: ^2.0.9
-  location: ^5.0.3
+  location: ^7.0.1
   badges: ^3.1.2
   dropdown_button2: ^2.3.9
   logger: ^2.4.0
 
   intl: ^0.19.0
   cherry_toast: ^1.11.0
-  change_app_package_name: ^1.3.0
-  firebase_core: ^3.5.0
-  firebase_messaging: ^15.1.2
-  firebase_analytics: ^11.3.2
+  change_app_package_name: ^1.4.0
+  firebase_core: ^3.6.0
+  firebase_messaging: ^15.1.3
+  firebase_analytics: ^11.3.3
   toast: ^0.3.0
-  url_launcher: ^6.3.0
+  url_launcher: ^6.3.1
   animated_custom_dropdown: ^3.1.1
 #  laravel_flutter_pusher_plus:
 #    git:
@@ -46,21 +46,21 @@ dependencies:
   cached_network_image: ^3.4.1
   flutter_image_compress: ^2.3.0
   flutter_widget_from_html_core: ^0.15.2
-  flutter_form_builder: ^9.4.1
+  flutter_form_builder: ^9.5.0
   form_builder_validators: ^11.0.0
 
   form_builder_image_picker: ^4.1.0
   form_builder_file_picker: ^4.2.0
   form_builder_extra_fields: ^11.0.0
-  share_plus: ^10.0.2
+  share_plus: ^10.1.1
   permission_handler: ^11.3.1
-  pusher_client_socket: ^0.0.1
+  pusher_client_socket: ^0.0.2+6
   app_links: ^6.3.2
   shimmer: ^3.0.0
   another_audio_recorder: ^1.2.1
   audioplayers: ^6.1.0
   youtube_player_flutter: ^9.1.1
-  lottie: ^3.1.2
+  lottie: ^3.1.3
   animated_icon: ^0.0.9
   google_sign_in: ^6.2.1
   auto_size_text: ^3.0.0
