@@ -1,16 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-
-import '../../helpers/colors.dart';
-import '../../helpers/style.dart';
 
 class RichEditor extends StatelessWidget {
    RichEditor({super.key,required this.controller});
 final TextEditingController controller;
   @override
   Widget build(BuildContext context) {
-    return Column(
+    return const Column(
       children: [
 
       ],
