@@ -1,13 +1,11 @@
-import 'package:ali_pasha_graph/Global/main_controller.dart';
-import 'package:ali_pasha_graph/components/advice_component/view.dart';
+
 import 'package:ali_pasha_graph/helpers/colors.dart';
 import 'package:ali_pasha_graph/helpers/style.dart';
-import 'package:ali_pasha_graph/models/product_model.dart';
-import 'package:cached_network_image/cached_network_image.dart';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:get/get.dart';
+
 import 'package:shimmer/shimmer.dart';
 
 class MinimizeDetailsProductComponentLoading extends StatelessWidget {
