@@ -1,24 +1,23 @@
-import 'package:ali_pasha_graph/Global/main_controller.dart';
-import 'package:ali_pasha_graph/models/setting_model.dart';
-import 'package:get/get.dart';
+
 import 'package:logger/logger.dart';
 import 'package:pusher_client_socket/pusher_client_socket.dart';
-// import 'package:laravel_flutter_pusher_plus/laravel_flutter_pusher_plus.dart';
 
 class PusherService {
 
   static PusherClient init({String? token}) {
 
     var options = PusherOptions(
-        host: '85.215.154.88:8081',
-        protocol:Protocol.ws,
+        host: 'pazarpasha.com',
+        // host: '85.215.154.88:8086',
+        // host: 'v3.ali-pasha.com:8086',
+       wsPort: 2053,
         autoConnect: false,
         cluster: 'mt1',
-        authOptions: PusherAuthOptions("http://v3.ali-pasha.com/api/broadcasting/auth",
+        authOptions: PusherAuthOptions("https://pazarpasha.com/api/broadcasting/auth",
             headers: {
               if (token!=null && token!='')
                 'Authorization': 'Bearer $token', // تمرير الـ token هنا
-              'Content-Type': 'application/json',
+             /* 'Content-Type': 'application/json',*/
             }), key: 'AliPasha',enableLogging: true,);
 
     final pusherClient = PusherClient( options: options);
@@ -29,6 +28,8 @@ logger.w(data);
     });
     pusherClient.onConnectionEstablished((data) {
       logger.w("Connection established - socket-id: ${pusherClient.socketId}");
+
+        // يمكنك استخدام Logger لطباعة الرسالة
     });
     pusherClient.onConnectionError((error) {
       logger.w("Connection error - $error");
