@@ -1,9 +1,7 @@
 import 'dart:io';
-import 'dart:math';
 
 import 'package:ali_pasha_graph/Global/main_controller.dart';
 import 'package:ali_pasha_graph/components/fields_components/input_component.dart';
-import 'package:ali_pasha_graph/components/product_components/job_card.dart';
 import 'package:ali_pasha_graph/components/progress_loading.dart';
 import 'package:ali_pasha_graph/helpers/colors.dart';
 import 'package:ali_pasha_graph/helpers/style.dart';
@@ -22,6 +20,7 @@ import 'package:image_cropper/image_cropper.dart';
 import 'package:image_picker/image_picker.dart';
 
 import '../../helpers/components.dart';
+import '../../helpers/helper_class.dart';
 import 'logic.dart';
 
 class EditProfilePage extends StatelessWidget {
@@ -105,9 +104,12 @@ class EditProfilePage extends StatelessWidget {
                     if (logic.user.value?.is_verified != true)
                       InkWell(
                           onTap: () {
-                            openUrl(
-                                url:
-                                    "https://wa.me/${mainController.settings.value.social?.phone}");
+                            HelperClass.requestVerified(onConfirm: (){
+                              if(isAuth()){
+                                String message="ID:${mainController.authUser.value?.id} - اسم المتجر : ${mainController.authUser.value?.seller_name} - نوع الطلب توثيق الحساب";
+                                openUrl(url: "https://wa.me/${mainController.settings.value.social?.phone}?text=$message");
+                              }
+                            });
                           },
                           child: Container(
                             width: 0.35.sw,
@@ -115,7 +117,7 @@ class EditProfilePage extends StatelessWidget {
                                 vertical: 0.01.sh, horizontal: 0.02.sw),
                             decoration: BoxDecoration(
                                 borderRadius: BorderRadius.circular(15.r),
-                                color: RedColor),
+                                color: Colors.blue),
                             child: Row(
                               mainAxisAlignment: MainAxisAlignment.center,
                               crossAxisAlignment: CrossAxisAlignment.center,
