@@ -1,11 +1,9 @@
-import 'dart:math';
 
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:form_builder_extra_fields/form_builder_extra_fields.dart';
+
 import 'package:form_builder_validators/form_builder_validators.dart';
 import 'package:get/get.dart';
 
@@ -14,7 +12,6 @@ import '../../helpers/colors.dart';
 import '../../helpers/style.dart';
 import '../../models/category_model.dart';
 import '../../models/city_model.dart';
-import '../../routes/routes_url.dart';
 import 'logic.dart';
 
 class EditServicePage extends StatelessWidget {
