@@ -76,7 +76,7 @@ class EditTenderLogic extends GetxController {
   }
 
   fillDataFromDraft() {
-    var data = box.read<Map<String, dynamic>>('draft');
+   // var data = box.read<Map<String, dynamic>>('draft');
   }
 
   Future<void> getDataForCreate() async {
